tests/wpt/tests/navigation-api/navigate-event/navigate-destination-getState-back-forward.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(t.step_func_done(() => {
    let navState = { statevar: "state" };
    navigation.navigate("#foo", { history: "replace", state: navState });
    let target_key = navigation.currentEntry.key;
    let target_id = navigation.currentEntry.id;
    navigation.navigate("#bar");
    navigation.onnavigate = t.step_func_done(e => {
      assert_equals(e.navigationType, "traverse");
      assert_not_equals(e.destination, null);
      assert_not_equals(e.destination.getState(), undefined);
      assert_not_equals(e.destination.getState(), e.destination.getState());
      assert_equals(e.destination.key, "");
      assert_equals(e.destination.id, "");
      assert_equals(e.destination.index, -1);
      assert_equals(e.sourceElement, null);
    });
    navigation.back();
  }), 0);
}, "navigate event destination.getState() on back/forward navigation");
</script>
