LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-destination-getState-back-forward-expected.txt


FAIL navigate event destination.getState() on back/forward navigation assert_equals: expected "traverse" but got "push"
